+++ lib/data/repositories/playlist_mutation_repository.dart
@@ -4,7 +4,7 @@ import 'package:fmp/core/logger.dart';
 import 'package:fmp/data/models/playlist.dart';
 import 'package:fmp/data/models/track.dart';
 import 'package:fmp/data/repositories/track_repository.dart';
-import 'package:fmp/services/library/playlist_exceptions.dart';
+import 'package:fmp/data/repositories/playlist_exceptions.dart';
 
 class PlaylistMutationResult {
   final int playlistId;
