+++ lib/data/sources/playlist_import/netease_playlist_source.dart
@@ -5,7 +5,7 @@ import 'package:fmp/i18n/strings.g.dart';
 import 'package:fmp/data/models/track.dart';
 import 'package:fmp/data/sources/source_http_policy.dart';
 import 'package:fmp/data/sources/source_url_policy.dart';
-import 'package:fmp/services/library/remote_playlist_id_parser.dart';
+import 'package:fmp/data/sources/remote_playlist_id_parser.dart';
 import 'package:fmp/data/sources/playlist_import/playlist_import_source.dart';
 
 /// 网易云音乐歌单导入源
