+++ lib/services/library/remote_playlist_sync_service.dart
@@ -1,5 +1,5 @@
 import 'package:fmp/data/models/playlist.dart';
-import 'package:fmp/services/library/remote_playlist_id_parser.dart';
+import 'package:fmp/data/sources/remote_playlist_id_parser.dart';
 
 class RemotePlaylistSyncService {
   final Future<List<Playlist>> Function() getImportedPlaylists;
